Remove broken option reset from per-file loop
When -loop is used as a per-file option it seeks back to the start of
the file from the end and tries to reset options to their original
values. This reset is not implemented properly: it simply sets the
option variables back without running any of the associated control
code. Implementing the option reset functionality properly would be a
significant amount of work and there is no clear need for it, so I'm
simply removing the broken version.

None of this affects the use of -loop as a global option. That case has
a separate implementation which always starts a new file from scratch
with default option values.