Fix configure regexp for enabling FFmpeg features
All features were disabled because of a regexp that used "[^ ]" to
match start of a word. '^' is not a metacharacter meaning beginning
of line inside [], instead it inverts the matching condition of the
other characters in the list.

Changed to use "\<" and "\>" for start/end of word instead. I don't know
how standard those are, at least they're better than the previous
version.