corevideo: Simplify mouse handling.
Handling of borders etc. should not be done inside the vo.
Neither the old nor the new version will be completely correct,
but the new version is simpler and will work correctly after
fixing a few other parts of the code/making -keepaspect behave
consistently between corevideo and other vos.